feat: remove button show register

+++ app/src/main/java/com/ceiba/parkinglot_adn/domain/objects/RegisterDomain.kt
@@ -1,3 +1,3 @@
 package com.ceiba.parkinglot_adn.domain.objects
 
-class RegisterDomain(plate: String, price: Double)
+class RegisterDomain(plate: String, price: Int)
